Expose MathJax option "tags" as a command-line option
This controls if and how labels and references are rendered.
Since the MathJax default is 'none', without a way to change this
option it is impossible to render labels with mjcli.

While the MathJax default is 'none', this commit sets the default
for mjcli to 'ams'.  The MathJax documentation states that the
'none' default has been chosen for backwards compatibility.
In practice, though, users most likely want support for labels.
For example, tex4ht in MathJax mode always uses 'aws'.